Move Deposit-related constants from NetworkConstants to pow.api.DepositConstants
