+++ beacon/pow/build.gradle
@@ -1,6 +1,5 @@
 dependencies {
   implementation project(':ethereum:pow:api')
-  implementation project(':ethereum:pow:merkletree')
   implementation project(':ethereum:spec')
   implementation project(':infrastructure:async')
   implementation project(':infrastructure:bls')
