+++ beacon/pow/src/test/java/tech/pegasys/teku/beacon/pow/Eth1DepositManagerTest.java
@@ -22,7 +22,6 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
-import static tech.pegasys.teku.ethereum.pow.merkletree.DepositTree.DEPOSIT_TREE_SNAPSHOT_SCHEMA;
 import static tech.pegasys.teku.infrastructure.async.SafeFuture.COMPLETE;
 
 import java.math.BigInteger;
@@ -386,7 +385,6 @@ class Eth1DepositManagerTest {
     final BigInteger lastBlockNumber = BigInteger.valueOf(1000);
     final DepositTreeSnapshot depositTreeSnapshot =
         new DepositTreeSnapshot(
-            DEPOSIT_TREE_SNAPSHOT_SCHEMA,
             List.of(dataStructureUtil.randomBytes32(), dataStructureUtil.randomBytes32()),
             dataStructureUtil.randomBytes32(),
             deposits.longValue(),
