+++ data/beaconrestapi/src/main/java/tech/pegasys/teku/beaconrestapi/handlers/tekuv1/beacon/GetDepositSnapshot.java
@@ -13,7 +13,6 @@
 
 package tech.pegasys.teku.beaconrestapi.handlers.tekuv1.beacon;
 
-import static tech.pegasys.teku.ethereum.pow.merkletree.DepositTree.DEPOSIT_TREE_SNAPSHOT_SCHEMA;
 import static tech.pegasys.teku.infrastructure.http.ContentTypes.JSON;
 import static tech.pegasys.teku.infrastructure.http.ContentTypes.OCTET_STREAM;
 import static tech.pegasys.teku.infrastructure.http.HttpStatusCodes.SC_NOT_FOUND;
@@ -51,7 +50,7 @@ public class GetDepositSnapshot extends MigratingEndpointAdapter {
   public static final String ROUTE = "/teku/v1/beacon/deposit_snapshot";
 
   private static final SerializableTypeDefinition<DepositTreeSnapshot> DEPOSIT_SNAPSHOT_TYPE =
-      DepositTreeSnapshot.getJsonTypeDefinition(DEPOSIT_TREE_SNAPSHOT_SCHEMA);
+      DepositTreeSnapshot.getJsonTypeDefinition();
 
   public static final SerializableTypeDefinition<DepositTreeSnapshot>
       DEPOSIT_SNAPSHOT_RESPONSE_TYPE =
