+++ ethereum/pow/api/src/main/java/tech/pegasys/teku/ethereum/pow/api/DepositTreeSnapshot.java
@@ -13,6 +13,8 @@
 
 package tech.pegasys.teku.ethereum.pow.api;
 
+import static tech.pegasys.teku.ethereum.pow.api.DepositConstants.DEPOSIT_CONTRACT_TREE_DEPTH;
+
 import java.util.List;
 import java.util.stream.Collectors;
 import org.apache.tuweni.bytes.Bytes;
@@ -20,8 +22,11 @@ import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.infrastructure.json.types.DeserializableTypeDefinition;
 import tech.pegasys.teku.infrastructure.ssz.SszList;
 import tech.pegasys.teku.infrastructure.ssz.containers.Container5;
+import tech.pegasys.teku.infrastructure.ssz.containers.ContainerSchema5;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszBytes32;
 import tech.pegasys.teku.infrastructure.ssz.primitive.SszUInt64;
+import tech.pegasys.teku.infrastructure.ssz.schema.SszListSchema;
+import tech.pegasys.teku.infrastructure.ssz.schema.SszPrimitiveSchemas;
 import tech.pegasys.teku.infrastructure.ssz.tree.TreeNode;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 
@@ -29,33 +34,60 @@ public class DepositTreeSnapshot
     extends Container5<
         DepositTreeSnapshot, SszList<SszBytes32>, SszBytes32, SszUInt64, SszBytes32, SszUInt64> {
 
-  public static DeserializableTypeDefinition<DepositTreeSnapshot> getJsonTypeDefinition(
-      final DepositTreeSnapshotSchema schema) {
-    return schema.getJsonTypeDefinition();
+  private static class DepositTreeSnapshotSchema
+      extends ContainerSchema5<
+          DepositTreeSnapshot, SszList<SszBytes32>, SszBytes32, SszUInt64, SszBytes32, SszUInt64> {
+
+    private DepositTreeSnapshotSchema() {
+      super(
+          "DepositTreeSnapshot",
+          namedSchema(
+              "finalized",
+              SszListSchema.create(
+                  SszPrimitiveSchemas.BYTES32_SCHEMA, DEPOSIT_CONTRACT_TREE_DEPTH)),
+          namedSchema("deposit_root", SszPrimitiveSchemas.BYTES32_SCHEMA),
+          namedSchema("deposit_count", SszPrimitiveSchemas.UINT64_SCHEMA),
+          namedSchema("execution_block_hash", SszPrimitiveSchemas.BYTES32_SCHEMA),
+          namedSchema("execution_block_height", SszPrimitiveSchemas.UINT64_SCHEMA));
+    }
+
+    @Override
+    public DepositTreeSnapshot createFromBackingNode(final TreeNode node) {
+      return new DepositTreeSnapshot(this, node);
+    }
+
+    @SuppressWarnings("unchecked")
+    public SszListSchema<SszBytes32, ? extends SszList<SszBytes32>> getFinalizedSchema() {
+      return (SszListSchema<SszBytes32, ? extends SszList<SszBytes32>>) getChildSchema(0);
+    }
   }
 
-  public static DepositTreeSnapshot fromBytes(
-      final Bytes bytes, final DepositTreeSnapshotSchema schema) {
-    return schema.sszDeserialize(bytes);
+  private static final DepositTreeSnapshotSchema SCHEMA = new DepositTreeSnapshotSchema();
+
+  public static DeserializableTypeDefinition<DepositTreeSnapshot> getJsonTypeDefinition() {
+    return SCHEMA.getJsonTypeDefinition();
+  }
+
+  public static DepositTreeSnapshot fromBytes(final Bytes bytes) {
+    return SCHEMA.sszDeserialize(bytes);
   }
 
   public DepositTreeSnapshot(
-      final DepositTreeSnapshotSchema schema,
       final List<Bytes32> finalized,
       final Bytes32 depositRoot,
       final long depositCount,
       final Bytes32 executionBlockHash,
       final UInt64 executionBlockHeight) {
     super(
-        schema,
-        finalized.stream().map(SszBytes32::of).collect(schema.getFinalizedSchema().collector()),
+        SCHEMA,
+        finalized.stream().map(SszBytes32::of).collect(SCHEMA.getFinalizedSchema().collector()),
         SszBytes32.of(depositRoot),
         SszUInt64.of(UInt64.valueOf(depositCount)),
         SszBytes32.of(executionBlockHash),
         SszUInt64.of(executionBlockHeight));
   }
 
-  protected DepositTreeSnapshot(final DepositTreeSnapshotSchema schema, final TreeNode node) {
+  private DepositTreeSnapshot(final DepositTreeSnapshotSchema schema, final TreeNode node) {
     super(schema, node);
   }
