+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/operations/Deposit.java
@@ -13,7 +13,7 @@
 
 package tech.pegasys.teku.spec.datastructures.operations;
 
-import static tech.pegasys.teku.spec.constants.NetworkConstants.DEPOSIT_CONTRACT_TREE_DEPTH;
+import static tech.pegasys.teku.ethereum.pow.api.DepositConstants.DEPOSIT_CONTRACT_TREE_DEPTH;
 
 import tech.pegasys.teku.infrastructure.ssz.collections.SszBytes32Vector;
 import tech.pegasys.teku.infrastructure.ssz.containers.Container2;
